Refactor: 사용자 API 정보 변경
